+++ main.cpp
@@ -69,9 +69,9 @@ int main() {
             cout << "Rozmiar > " << MAX_BF_SIZE << " i brak 'sum_min'. Pomijam wyznaczanie optimum." << endl;
         }
 
-        // Determine initial Upper Bound for Branch and Bound
-        long long initialUB = numeric_limits<long long>::max();
-        if (cfg.algorithm.find("BB") == 0) {
+         // Determine initial Upper Bound for Branch and Bound
+         int initialUB = numeric_limits<int>::max();
+         if (cfg.algorithm.find("BB") == 0) {
             cout << "Wyznaczanie poczatkowego Gornego Ograniczenia (UB) - strategia: " << cfg.ubStrategy << endl;
 
             if (cfg.ubStrategy == "RAND") {
